feat: renamed all file to integration-pills, added cn
