+++ content/docs/card/integration-pills.mdx
@@ -4,7 +4,7 @@ description: A component that displays a list of services it can integrate with.
 author: shoomankhatri
 ---
 
-<ComponentPreview name="card-feature-card-tsx--docs" />
+<ComponentPreview name="card-integration-pills--docs" />
 
 ## Installation
 
@@ -12,17 +12,17 @@ author: shoomankhatri
 
 <Step>Run the following command</Step>
 
-It will create a new file `feature-card-tsx.tsx` inside the `components/animata/card` directory.
+It will create a new file `integration-pills.tsx` inside the `components/animata/card` directory.
 
 ```bash
-mkdir -p components/animata/card && touch components/animata/card/feature-card-tsx.tsx
+mkdir -p components/animata/card && touch components/animata/card/integration-pills.tsx
 ```
 
 <Step>Paste the code</Step>{" "}
 
 Open the newly created file and paste the following code:
 
-```jsx file=<rootDir>/animata/card/feature-card-tsx.tsx
+```jsx file=<rootDir>/animata/card/integration-pills.tsx
 
 ```
